Crate synthpop

This crate describes a synthetic population that exist in a map. Currently each person’s travel behavior is modelled, but in the future, demographic and health attributes may be added. There’s a variety of ways to create these populations, scattered in other crates.

Note that “scenario” is the term currently used to describe the population. This will be renamed “soon.”

Modules

borders

counts

endpoint

external

Some users of the API (https://a-b-street.github.io/docs/tech/dev/api.html) have their own simulation input data; import it here.

modifier

scenario

Structs

ExternalPerson

ExternalTrip

IndividTrip

MapBorder

MapBorders

Lists all border intersections of the map, broken down by mode and whether they support incoming or outgoing traffic.

OrigPersonID

This is an ID used by Seattle soundcast. Originally it was preserved for debugging, but that hasn’t happened in a long time. Also the format is tied to Soundcast. Consider deleting / changing.

PersonSpec

Scenario

A Scenario describes all the input to a simulation. Usually a scenario covers one day.

TrafficCounts

This represents the number of vehicles (or trips, or something else) crossing roads and intersections over some span of time. The data could represent real observations or something from a simulation.

Enums

ExternalTripEndpoint

ScenarioModifier

Transforms an existing Scenario before instantiating it.

TripEndpoint

Specifies where a trip begins or ends.

TripMode

How does a trip primarily happen?

TripPurpose

Lifted from Seattle’s Soundcast model, but seems general enough to use anyhere.